llm-test/requirements.txt
llama-cpp-python
gradio
huggingface_hub
colorama 